+++ src/TokenStorage.js
@@ -2,7 +2,6 @@ class TokenStorage {
   constructor(tokenGenerator, asyncStorage, accessTokenKey = 'rest_client_sdk.api.access_token') {
     this._tokenGenerator = tokenGenerator;
     this._hasATokenBeenGenerated = false;
-    this._tokenExpiresAtMap = {};
     this.setAsyncStorage(asyncStorage);
     this.accessTokenKey = accessTokenKey;
   }
@@ -53,7 +52,7 @@ class TokenStorage {
 
   generateToken(parameters) {
     this._hasATokenBeenGenerated = true;
-    const callTimestamp = Date.now();
+    const callTimestamp = Date.now() / 1000;
     return this._tokenGenerator
       .generateToken(parameters)
       .then(responseData =>
@@ -65,7 +64,7 @@ class TokenStorage {
     return this._asyncStorage
       .getItem(this.accessTokenKey)
       .then(token => {
-        const callTimestamp = Date.now();
+        const callTimestamp = Date.now() / 1000;
         return this._tokenGenerator
           .refreshToken(JSON.parse(token), parameters)
           .then(responseData =>
@@ -79,21 +78,32 @@ class TokenStorage {
    * Return the number of second when the token will expire
    * return value can be negative if the token is already expired
   */
-  getTokenExpiresIn(token) {
-    return this._tokenExpiresAtMap[token.access_token] - (Date.now() / 1000);
+  getCurrentTokenExpiresIn() {
+    return this.getAccessTokenObject()
+    .then(accessTokenObject => {
+      if (accessTokenObject === null || typeof accessTokenObject.expires_at === 'undefined') {
+        throw new Error('No token has been stored.');
+      }
+
+      const now = Date.now() / 1000;
+
+      return accessTokenObject.expires_at - now;
+    });
   }
 
   _storeAccessToken(responseData, callTimestamp) {
+    let responseDataToStore = responseData;
     if (typeof responseData === 'object') {
-      this._tokenExpiresAtMap[responseData.access_token] = null;
+      responseDataToStore = Object.assign({}, responseData);
+      responseDataToStore.expires_at = null;
       if (typeof responseData.expires_in !== 'undefined' && responseData.expires_in >= 0) {
-        this._tokenExpiresAtMap[responseData.access_token] = (Date.now() / 1000) + responseData.expires_in;
+        responseDataToStore.expires_at = callTimestamp + responseData.expires_in;
       }
     }
 
     return this._asyncStorage.setItem(
       this.accessTokenKey,
-      JSON.stringify(responseData)
+      JSON.stringify(responseDataToStore)
     );
   }
 }